ccleanup/split of McpQueryPage (#6758)
